refactor: Level.1 - 4. 관심사 분리
JWT 처리 로직과 비즈니스 로직을 분리하고, 인증된 사용자 정보를 더 적절한 방식으로 가져올 수 있도록 deleteManager() 메서드를 개선

+++ src/main/java/org/example/expert/domain/manager/controller/ManagerController.java
@@ -1,11 +1,14 @@
 package org.example.expert.domain.manager.controller;
 
 import io.jsonwebtoken.Claims;
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.apache.coyote.BadRequestException;
 import org.example.expert.config.JwtUtil;
 import org.example.expert.domain.common.annotation.Auth;
 import org.example.expert.domain.common.dto.AuthUser;
+import org.example.expert.domain.common.exception.InvalidRequestException;
 import org.example.expert.domain.manager.dto.request.ManagerSaveRequest;
 import org.example.expert.domain.manager.dto.response.ManagerResponse;
 import org.example.expert.domain.manager.dto.response.ManagerSaveResponse;
@@ -38,12 +41,15 @@ public class ManagerController {
 
     @DeleteMapping("/todos/{todoId}/managers/{managerId}")
     public void deleteManager(
-            @RequestHeader("Authorization") String bearerToken,
+            HttpServletRequest request,
             @PathVariable long todoId,
             @PathVariable long managerId
     ) {
-        Claims claims = jwtUtil.extractClaims(bearerToken.substring(7));
-        long userId = Long.parseLong(claims.getSubject());
+        Long userId = (Long) request.getAttribute("userId");    // HttpServletRequest에서 userId 얻기
+        // userId가 없을 경우
+        if (userId == null) {
+            throw new InvalidRequestException("인증된 사용자 정보가 없습니다.");
+        }
         managerService.deleteManager(userId, todoId, managerId);
     }
 }
